backend/requirements.txt
# Backend Dependencies

Flask==3.1.0
Flask-SocketIO==5.5.1  # Latest version for better WebSocket support
pymongo==4.11.3  # MongoDB connection
pyjwt==2.9.0  # JWT authentication
websocket-client==1.7.0  # Correct WebSocket dependency for Flask-SocketIO
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
Werkzeug==3.1.3

# Utility & Security
bidict==0.23.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
wsproto==1.2.0

# Frontend Dependencies (PyQt6/PySide6)
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
shiboken6==6.8.2.1
